Clean up, simplify, use autoloading, use more closures to simplify further, add namespacing usage; add helpful constants
