+++ includes/customizer.php
@@ -1,8 +1,6 @@
 <?php
 /**
  * Heisenberg Theme Customizer.
- *
- * @package Heisenberg
  */
 
 /**
@@ -10,17 +8,21 @@
  *
  * @param WP_Customize_Manager $wp_customize Theme Customizer object.
  */
-function heisenberg_customize_register( $wp_customize ) {
+add_action( 'customize_register', function( $wp_customize ) {
 	$wp_customize->get_setting( 'blogname' )->transport         = 'postMessage';
 	$wp_customize->get_setting( 'blogdescription' )->transport  = 'postMessage';
 	$wp_customize->get_setting( 'header_textcolor' )->transport = 'postMessage';
-}
-add_action( 'customize_register', 'heisenberg_customize_register' );
+} );
 
 /**
  * Binds JS handlers to make Theme Customizer preview reload changes asynchronously.
  */
-function heisenberg_customize_preview_js() {
-	wp_enqueue_script( 'heisenberg_customizer', get_template_directory_uri().'/inc/js/customizer.js', array( 'customize-preview' ), '20130508', true );
-}
-add_action( 'customize_preview_init', 'heisenberg_customize_preview_js' );
+add_action( 'customize_preview_init', function() {
+	wp_enqueue_script(
+		'heisenberg_customizer',
+		HEISENBERG_URL . '/assets/js/customizer.js',
+		['customize-preview'],
+		HEISENBERG_VERSION,
+		true
+	);
+} );
