+++ includes/jetpack.php
@@ -17,7 +17,7 @@ function heisenberg_jetpack_setup() {
 		'render'    => 'heisenberg_infinite_scroll_render',
 		'footer'    => 'page',
 	) );
-} // end function heisenberg_jetpack_setup
+}
 add_action( 'after_setup_theme', 'heisenberg_jetpack_setup' );
 
 /**
@@ -28,4 +28,4 @@ function heisenberg_infinite_scroll_render() {
 		the_post();
 		get_template_part( 'page-templates/content', get_post_format() );
 	}
-} // end function heisenberg_infinite_scroll_render
+}
