+++ package.json
@@ -1,24 +1,7 @@
 {
-  "devDependencies": {
-    "babel-preset-es2015": "^6.9.0",
-    "browser-sync": "^2.12.5",
-    "del": "^2.2.1",
-    "gulp": "^3.9.0",
-    "gulp-autoprefixer": "^3.1.0",
-    "gulp-babel": "^6.1.2",
-    "gulp-concat": "^2.6.0",
-    "gulp-jshint": "^1.11.0",
-    "gulp-notify": "^2.2.0",
-    "gulp-sass": "^2.3.1",
-    "gulp-size": "^2.1.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-svg-sprite": "^1.2.19",
-    "gulp-svgmin": "^1.2.2",
-    "gulp-uglify": "^1.2.0"
-  },
   "name": "heisenberg",
   "description": "Heisenberg - Zeek Starter Theme ===",
-  "version": "2.0.2",
+  "version": "2.0.3",
   "main": "gulpfile.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -35,5 +18,22 @@
   "homepage": "https://github.com/ZeekInteractive/heisenberg",
   "dependencies": {
     "foundation-sites": "^6.2.3"
+  },
+  "devDependencies": {
+    "babel-preset-es2015": "^6.9.0",
+    "browser-sync": "^2.12.5",
+    "del": "^2.2.1",
+    "gulp": "^3.9.0",
+    "gulp-autoprefixer": "^3.1.0",
+    "gulp-babel": "^6.1.2",
+    "gulp-concat": "^2.6.0",
+    "gulp-jshint": "^1.11.0",
+    "gulp-notify": "^2.2.0",
+    "gulp-sass": "^2.3.1",
+    "gulp-size": "^2.1.0",
+    "gulp-sourcemaps": "^1.6.0",
+    "gulp-svg-sprite": "^1.2.19",
+    "gulp-svgmin": "^1.2.2",
+    "gulp-uglify": "^1.2.0"
   }
 }
