+++ style.css
@@ -4,7 +4,7 @@ Theme URI: https://github.com/ZeekInteractive/heisenberg
 Author: Zeek Interactive
 Author URI: http://zeek.com
 Description: Description
-Version: 2.0.1
+Version: 2.0.3
 License: GNU General Public License v2 or later
 License URI: http://www.gnu.org/licenses/gpl-2.0.html
 Text Domain: heisenberg
